+++ Runtime/Adapters/Vector4Adapter.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace Aceland.KalmanFilter.Adapters
 {
     [BurstCompile]
-    internal struct Vector4Adapter : IKalmanValueAdapter<Vector4>
+    public readonly struct Vector4Adapter : IKalmanValueAdapter<Vector4>
     {
         public Vector4 Zero => Vector4.zero;
         public Vector4 Add(Vector4 left, Vector4 right) => left + right;